+++ nbproject/private/private.xml
@@ -3,12 +3,7 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/E:/SWP291_Group4_ProjectMedicine-haidt/src/java/controller/ImportList.java</file>
-            <file>file:/E:/SWP291_Group4_ProjectMedicine-haidt/src/java/model/Unit.java</file>
-            <file>file:/E:/SWP291_Group4_ProjectMedicine-haidt/web/importList.jsp</file>
-            <file>file:/E:/SWP291_Group4_ProjectMedicine-haidt/src/java/dal/stockDAO.java</file>
-            <file>file:/E:/SWP291_Group4_ProjectMedicine-haidt/src/java/controller/importServlet.java</file>
-            <file>file:/E:/SWP291_Group4_ProjectMedicine-haidt/web/importProduct.jsp</file>
+            <file>file:/E:/SWP291_Group4_ProjectMedicine-haidt/src/java/controller/CategoryServlet.java</file>
         </group>
     </open-files>
 </project-private>
